gmnlm: make ./ prefix mandatory for relative URLs
Interpreting everything that isn't a command as relative URL wasn't
very intuitive and lead to accidental commands or requests being made.
This commit aims to clearly separate relative URLs from commands, which
also allows us to add basic error reporting for commands.

The only exception that had to be made was in regards to /whatever URLs
as those are impossible to support without changing the search command.